Workaround element/attribute names with special characters
For element names, these are equivalent in XPath:

    foo
    *[name() = "foo"]

And for attribute names:

    @foo
    attribute:*[name() = "foo"]

The former is faster but some characters are not allowed in it.
Since I am not sure which characters, only use it for "safe" names
that match ^[a-zA-Z_][a-zA-Z0-9_.-]*$
This is overly restrictive, but should cover every name actually used
in XML, HTML, SVG, etc.